Add RSAEP and RSADP

+++ includes/std.h
@@ -74,7 +74,7 @@ struct __attribute__((__packed__)) rsa {
 /* process.c */
 uint64_t			custom_rand(void);
 uint64_t			rand_range(uint64_t min, uint64_t max);
-uint64_t			power_mod(uint64_t x, uint64_t n, uint64_t p);
+unsigned __int128	power_mod(unsigned __int128 x, unsigned __int128 n, unsigned __int128 p);
 unsigned __int128	pgcd_binary(unsigned __int128 a, unsigned __int128 b);
 unsigned __int128	inv_mod(unsigned __int128 a, unsigned __int128 n);
 bool				check_prime(uint64_t n, double proba, bool verbose);
